requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
opencc-python-reimplemented>=0.1.7
emoji>=2.0.0
markdown>=3.0.0
